refactor(LottoRun): 불필요한 호출 개선
- 불필요하게 LottoRun의 메소드를 통해 WinLotto의 메소드를 호출
- LottoRun의 메소드 제거, WinLotto 메소드 직접 호출
